requirements.txt
--find-links https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.24.0
tqdm>=4.62.0
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
nltk>=3.8.1 